refactor: 각 컴포넌트의 index  파일명 변경(코드컨벤션)
