+++ src/components/common/index.ts
@@ -1,2 +1,2 @@
-export { default as Header } from './Header';
-export { default as TabBar } from './TabBar';
+export { default as Header } from './Header/Header';
+export { default as TabBar } from './TabBar/TabBar';
