+++ src/components/index.ts
@@ -1,1 +1,1 @@
-export { default as DefaultTemplate } from './DefaultTemplate';
+export { default as DefaultTemplate } from './DefaultTemplate/DefaultTemplate';
